add adapted version of commit b760284
namely the one introducing the variable `deft-pending-updates'

I.e., this one:

Author: Jason Blevins <jrblevin@sdf.org>
Date:   Fri Apr 22 12:44:53 2016 -0400

Lazily update the Deft buffer after saving files

When a Deft note is saved, if the Deft buffer is not visible, then
postpone the (potentially costly) filter refresh until later.  This
reduces the number of such background updates that occur while editing
files, especially when the files are being saved automatically.