remove `notdeft-ensure-init'
It gets on the way having to have it at most entry points.
Instead, favor on-load initialization.

To facilitate this change, assume that
`notdeft-directories' has been set for any entry.

To still support dynamic recomputation of
`notdeft-directories', add a `notdeft-path' feature
for managing that. However, simplify that functionality
so that `notdeft-path' is either a list or function:

  - that is, a function returning a list of strings,
    for simplicity, rather than a safe to set buffer
    locally DSL, as `notdeft-directories' is not
    buffer local

Also introduce a `rescan' pending state, which is set
initially, so that any initial refresh of NotDeft buffer
window causes sufficient work to be done to still
get the file list displayed.