optimize `notdeft-changed/query'
Have it not update `notdeft-all-files', except indirectly when a NotDeft
buffer window is visible.

This results in clearer overall semantics for `notdeft-changed', such
that none of them touch the all-files or current-files lists, unless a
buffer to view them is there. Any Xapian indexes are still kept up to
date, and the pending flag will record any need to possibly later update
the file list state.