parse in-buffer setting keywords ignoring case
So that e.g. "#+TITLE" and "#+title" are treated the same.

That was actually already the case for `notdeft-parse-buffer',
as case folding regexp matching is the default in Emacs.
Thus the parsing behavior has been inconsistent between
Emacs Lisp and the C++ program so far.

This change to search indexing should remove that inconsistency.

Using lowercase keywords is apparently even preferred now in Org, as
pointed out by Kaushal Modi, so it's appropriate to support them.