add support for sparse dirs to the C++ program
That is, for the "index" command, allow an explicit file
listing for a directory to be indexed, so that nothing
else is considered to be a note file. Filename extensions
are ignored in this case. Any file not listed is removed
from the search index.

To support such file listings, add support for reading
indexing instructions from STDIN, when given the --input
option. This is to avoid running into maximum command
line lengths. The supported instructions are
- ":idir" instruction indexes a full dir tree
- ":ifiles" instruction indexes specified files
and any number of these may be given.