add "path:" prefix support
For matching against the directory path strings of notes.

This may be useful in cases where files have already been
organized into separate directories according to their subject
matter. It may then be redundant to tag the files, if a "path:"
search will do for narrowing the search down to a given topic.