ask for confirmation when moving subdirectories
When called interactively, that is, for convenience. Still allow a
prefix argument as another form of confirmation.

Also, handle the case where note(s) are open from a subdirectory that
gets moved or archived, meaning that the state of those buffers is
perhaps best altered also, so that they get written at the new
destination on save. (Exclude non-NotDeft buffers from that buffer file
name modifying behavior, although it is also possible that some note
attachments might be open when moving a subdirectory. Such attachments
would probably tend to be read only, and thus not to be edited or saved
in any case.)