modify Org linking functions
Support the use of an active region as a basis of a link
description, and also the replacing of a region with a link,
which tends to be useful for a text that has already been
written, and which one wants to link to other notes.

The two functions are now named
- `notdeft-org-link-existing-note'
- `notdeft-org-link-new-file'

Also, remove `notdeft-kill-ring-save-org-link'.